working extract-tables for multi-statement inputs
using an unholy combination of sqlparse and sqlglot.

sqlparse is more lenient and can chop up multiple statements even though
it doesn't understand "\T".  sqlglot is stricter, and chokes when it
sees "\T", but is better at the simple task of extracting table names
from a valid statement, and provides a much better interface.

We could also use sqlglot.tokenize() and split on the semicolon token,
which would be more verbose.

Closes #1122, in which the table name was always given as "DUAL" when
multiple statements were in the same input leading with "\T sql-insert".