Trailing shell-style redirect syntax
Inspired by redis-cli, add trailing shell-style redirect syntax.  This
mostly duplicates \once and \pipe_once in functionality but with much
greater convenience.

Before:

    \T csv;
    \once -o name.csv;
    select * from user where username = 'name';
    \T ascii;

After

    select * from user where username = 'name' $> name.csv;

Like the shell, a single angle bracket overwrites, and a double angle
bracket appends. The "$" character is needed to resolve ambiguity with
SQL's built-in operators.

Limitation: the filename must not contain spaces.

Limitation: a trailing semicolon is still required in multiline mode.

Limitation: the input line must be a single statement.

The algorithm is:
 * does the input contain a dollar-operator sequence, as tokenized by
   sqlglot?  This guarantees that the sequence is outside of a quote.
 * is the text to the left of the rightmost dollar-operator valid SQL?
 * is the text to the right of the rightmost dollar-operator free of
   space characters?
if so, execute the SQL to the left of the operator and store the results
the file named after the right of the operator, appending or
overwriting depending on the value of the operator.

Limitation: if any of the above conditions is not met, mycli will
fall back to dispatching the command as plain SQL.  This means that
error messages may not be clear when the user's intent is to use
redirection, but the input is somehow not parseable.

A new setting \Tr, alias redirectformat, is introduced, making the
redirected output format independent from the main interactive format.
This defaults to CSV, which can be considered a breaking change, since
previously the format was whatever the user had selected for
interactive use.  The default for redirectformat can also be changed in
myclirc.

\once, \pipe_once, and shell-style redirects now redirect content only
to the designated output target and not to the TUI.  This can also be
considered a breaking change.

tee is unchanged here, as the intention with tee may be different,
considering that queries are also logged.

Support for a "|" operator is also included.  In this case, the shell
portion may contain spaces.  The limitation is that there may be only
one "|", not a chained pipeline.  Still, one can do

    select * from user where username = 'name' $| gh gist create;

or

    select 100 as constant $| jq '. | length';
    8
    3