use plain print() to communicate with subprocess
Some lockups were observed when redirecting large outputs to "jq".
click.echo does file.write() and file.flush() which is probably not
what we want given that line-buffering was requested in Popen().

We might also want to avoid setting these:

    stdout=subprocess.PIPE,
    stderr=subprocess.PIPE,

and just allow the subprocess to communicate to the TTY.  That's where
the output goes eventually anyway.