support chained pipe operations
 * recast variables from parse as sql_part, operator_part, and
   shell_part.  Use these consistently.
 * cache get_redirect_command() since it is called twice: once to see
   if a redirected is wanted; the second time for the parts.
 * break redirect parser into units.
 * rewrite parse to allow chained pipe operations, accepting any number
   of "$|".  Disallow chained "$|" on Windows, since there isn't a POSIX
   sh we can depend on.  pipe_once may already not work well on Windows.

The parse rule is changed from looking for the rightmost matching shell
operator to the leftmost matching operator.  This can have no effect
for file redirections, since the angle bracket was and is a disallowed
character in the file part.

A chain of "$|" followed by "$>" is disallowed in the parse, but there
is a comment that this is a reasonable future feature.

This implementation currently has the bug/hidden feature that only the
leftmost "$|" must have the dollar sign.  That probably must be fixed.

The behavior of \pipe_once is changing on certain errors.  The error is
not thrown immediately for example on "\pipe_once nonexistent_command".
It is not thrown until some input is given.  This is required by the
switch to communicate(), and is much more predictable behavior, as
previously only some types of errors were frontloaded.