make control-r reverse search style configurable
Allow the old control-r incremental search to be recovered via a config
file setting.  In case control-r is bound to traditional incremental
search, add alt-r bindings for fzf search so that it is always
available.

The "fzf" setting for control_r in myclirc remains a bit of a lie,
because we still fall back to reverse_isearch if fzf is not installed.
But it could be nice to emit a warning in that case.

This is intended to address the discussion comment at

 * https://github.com/dbcli/mycli/discussions/1265#discussioncomment-13791547