fix traditional repeated reverse isearch
when not using fzf.  According to

 * https://github.com/prompt-toolkit/python-prompt-toolkit/blob/3374ae9dd56f0265b41022e81341fd062092e2f0/src/prompt_toolkit/key_binding/bindings/search.py#L47-L53

it looks like we need to set the "control_is_searchable" filter in
order to map control-r (and escape-r) only when _outside_ the isearch
mini-mode keymap.

It isn't clear how to specify both emacs_mode and control_is_searchable
as filters, so emacs_mode was removed.

Closes #1310.